requirements.txt
streamlit==1.25.0
beautifulsoup4==4.12.3
requests==2.31.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyMuPDF==1.23.7
scikit-learn==1.3.2
pandas==2.2.2
numpy==1.24.3
huggingface_hub==0.4.0
